Fix capitalisation.
The question whether to write compound words in titles with only the
first part capitalised or all of them capitalised (e.g. "Non-functional
Requirements" vs "Non-Functional Requirements") is a bit debatable,
especially in American English, however given one such word appears in
the title and your abbreviation for that title does not include the
second M in "in-memory", I have gone ahead and applied the former form
everywhere for consistency.