UPD: MSP: MSPConsumer: refine descriptions.
Remark:
We already highlight an interface as being an interface.
The reader of the document should understand that an interface is a
method set to be implemented by a particular struct.
No need to re-explain this in the beginning of every interface
description.

Same goes for similar things (struct, method, ...)